Rename package delegated_credentials to delegated_credential
